While playing a game, iOS system events can cause inconsistent input detection.
This example is used to demonstrate input stutter caused by iOS system
events and fluctuating frame rate.

While the game is running, access the Command Center and Notification
Center by swiping from the top and bottom edge of the screen. This
either causes the touch input to update at a slower rate than the game
is running, or restore the touch input to the same rate as the game is
running.